+++ lib/active_remote/search.rb
@@ -22,7 +22,7 @@ module ActiveRemote
       #   Tag.find(Tag.new(:guid => 'foo'))
       #
       #   # Protobuf object
-      #   Tag.find(Generic::Remote::TagRequest.new(:guid => 'foo'))
+      #   Tag.find(Generic::Remote::TagRequest.new(:guid => ['foo']))
       #
       def find(args)
         remote = search(args).first
@@ -42,7 +42,7 @@ module ActiveRemote
       #   Tag.find_by(Tag.new(:guid => 'foo'))
       #
       #   # Protobuf object
-      #   Tag.find_by(Generic::Remote::TagRequest.new(:guid => 'foo'))
+      #   Tag.find_by(Generic::Remote::TagRequest.new(:guid => ['foo']))
       #
       def find_by(args)
         search(args).first
@@ -57,21 +57,19 @@ module ActiveRemote
       #   Tag.first_or_create(:name => 'foo')
       #
       #   # Protobuf object
-      #   Tag.first_or_create(Generic::Remote::TagRequest.new(:name => 'foo'))
+      #   Tag.first_or_create(Generic::Remote::TagRequest.new(:name => ['foo']))
       #
       def first_or_create(attributes)
-        remote = search(attributes).first
-        remote ||= create(attributes)
-        remote
+        attributes = validate_search_args!(attributes)
+        search(attributes).first || create(attributes_for_record(attributes))
       end
 
       # Tries to load the first record; if it fails, then create! is called
       # with the same arguments.
       #
       def first_or_create!(attributes)
-        remote = search(attributes).first
-        remote ||= create!(attributes)
-        remote
+        attributes = validate_search_args!(attributes)
+        search(attributes).first || create!(attributes_for_record(attributes))
       end
 
       # Tries to load the first record; if it fails, then a new record is
@@ -83,12 +81,11 @@ module ActiveRemote
       #   Tag.first_or_initialize(:name => 'foo')
       #
       #   # Protobuf object
-      #   Tag.first_or_initialize(Generic::Remote::TagRequest.new(:name => 'foo'))
+      #   Tag.first_or_initialize(Generic::Remote::TagRequest.new(:name => ['foo']))
       #
       def first_or_initialize(attributes)
-        remote = search(attributes).first
-        remote ||= new(attributes)
-        remote
+        attributes = validate_search_args!(attributes)
+        search(attributes).first || new(attributes_for_record(attributes))
       end
 
       # Searches for records with the given arguments. Returns a collection of
@@ -100,7 +97,7 @@ module ActiveRemote
       #   Tag.search(:name => 'foo')
       #
       #   # Protobuf object
-      #   Tag.search(Generic::Remote::TagRequest.new(:name => 'foo'))
+      #   Tag.search(Generic::Remote::TagRequest.new(:name => ['foo']))
       #
       def search(args)
         args = validate_search_args!(args)
@@ -118,15 +115,32 @@ module ActiveRemote
       # Search args must be a hash or respond to to_hash
       #
       def validate_search_args!(args)
-        unless args.is_a?(Hash)
-          if args.respond_to?(:to_hash)
-            args = args.to_hash
-          else
-            raise "Invalid parameter: #{args}. Search args must respond to :to_hash."
-          end
-        end
+        return args if args.is_a?(Hash)
+        return args.attributes if args.is_a?(::ActiveRemote::Base)
+        return args.to_hash if args.respond_to?(:to_hash)
 
-        args
+        raise "Invalid parameter: #{args}. Search args must respond to :to_hash."
+      end
+
+      private
+
+      # Search fields are repeated so callers can match many records at once,
+      # but the record's attribute is scalar: ["foo"] would cast to "[\"foo\"]".
+      #
+      def attributes_for_record(args)
+        args.to_h do |name, value|
+          next [name, value] unless value.is_a?(::Array)
+
+          # A type that takes the array unchanged is meant to hold it.
+          next [name, value] if attribute_types[name.to_s].cast(value) == value
+
+          if value.size > 1
+            raise ArgumentError, "Cannot build #{self} from #{name.inspect} => #{value.inspect}. " \
+              "#{name} holds a single value, but #{value.size} were given."
+          end
+
+          [name, value.first]
+        end
       end
     end
 
@@ -135,6 +149,7 @@ module ActiveRemote
     def reload
       fresh_object = self.class.find(scope_key_hash)
       @attributes = fresh_object.instance_variable_get(:@attributes)
+      @new_record = false
       self
     end
   end